rename Backend + InMemoryBackend to State + Multiproof
