+++ src/hash.rs
@@ -1,5 +1,7 @@
 use sha2::{Digest, Sha256};
 
+pub type H256 = [u8; 32];
+
 pub fn hash(buf: &mut [u8; 64]) {
     let mut tmp = [0u8; 32];
     tmp.copy_from_slice(Sha256::digest(buf).as_ref());
